Moved log out button component to src/components
